% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expect.R
\name{expect_lazyforward}
\alias{expect_lazyforward}
\title{Test NSE-SE function pairs}
\usage{
expect_lazyforward(nse, se = paste0(nse, "_"), env = parent.frame(),
  .dots = ".dots")
}
\description{
Allows simple testing of hand-written NSE-SE pairs.
}
\seealso{
\code{\link{lazyforward}}
}

